move get_number_of_args to own file, write a py3-specific version
I've also re-worked the Python 2 version so that it should now work for Python 3.

But the Python-3 only version is much simpler. Also, the Python 2 version raises deprecation warnings in Python 3 (for getargspec).
